requirements.txt
asgiref==3.7.2
boto3==1.28.61
botocore==1.31.61
certifi==2023.7.22
charset-normalizer==3.3.0
Django==3.2.21
django-storages==1.14.1
djangorestframework==3.14.0
easy-colors==0.0.1
idna==3.4
itsdangerous==2.1.2
jmespath==1.0.1
Pillow
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-resize-image==1.1.20
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.17
whitenoise==6.5.0
gunicorn==21.2.0
django-cors-headers
dj-database-url==2.1.0
python-dotenv
drf-yasg
